port token saves from base.p8

+++ temple.p8
@@ -24,13 +24,13 @@ end
 
 
 objects,got_fruit, --tables
-freeze,delay_restart,sfx_timer,music_timer,ui_timer, --timers
+freeze,delay_restart,sfx_timer,ui_timer, --timers
 cam_x,cam_y,cam_spdx,cam_spdy,cam_gain,cam_offx,cam_offy, --camera values <camtrigger>
 _pal, --for outlining
 shake,screenshake
 =
 {},{},
-0,0,0,0,-99,
+0,0,0,-99,
 0,0,0,0,0.1,0,0,
 pal,
 0,false
@@ -41,7 +41,7 @@ local _g=_ENV --for writing to global vars
 -- [entry point]
 
 function _init()
-  max_djump,deaths,frames,seconds,minutes,music_timer,time_ticking,berry_count=1,0,0,0,0,0,true,0
+  max_djump,deaths,frames,seconds,minutes,time_ticking,berry_count=1,0,0,0,0,true,0
   music(0,0,7)
   load_level(1)
 end
@@ -49,29 +49,25 @@ end
 
 -- [effects]
 
-function rnd128()
-  return rnd(128)
-end
-
 clouds={}
 for i=0,16 do
   add(clouds,{
-    x=rnd128(),
-    y=rnd128(),
-    spd=1+rnd(4),
-    w=32+rnd(32)
+    x=rnd"128",
+    y=rnd"128",
+    spd=1+rnd"4",
+    w=32+rnd"32"
   })
 end
 
 particles={}
 for i=0,24 do
   add(particles,{
-    x=rnd128(),
-    y=rnd128(),
-    s=flr(rnd(1.25)),
-    spd=0.25+rnd(5),
+    x=rnd"128",
+    y=rnd"128",
+    s=flr(rnd"1.25"),
+    spd=0.25+rnd"5",
     off=rnd(),
-    c=6+rnd(2),
+    c=6+rnd"2",
   })
 end
 
@@ -85,11 +81,9 @@ player={
 
     --<fruitrain>--
     -- ^ refers to setting berry_timer and berry_count to 0
-    for var in all(split"grace,jbuffer,dash_time,dash_effect_time,\z
-                         dash_target_x,dash_target_y,dash_accel_x,dash_accel_y,\z
-                         spr_off,berry_timer,berry_count") do
+    foreach(split"grace,jbuffer,dash_time,dash_effect_time,dash_target_x,dash_target_y,dash_accel_x,dash_accel_y,spr_off,berry_timer,berry_count", function(var)
       _ENV[var]=0
-    end
+    end)
     -- <keydoor> --
     key_count=0
     -- </keydoor> --
@@ -101,7 +95,7 @@ player={
     end
 
     -- horizontal input
-    local h_input=btn(➡️) and 1 or btn(⬅️) and -1 or 0
+    local h_input=split"0,-1,1,1"[btn()%4+1]
 
     -- spike collision / bottom death
     if is_flag(0,0,-1) or
@@ -119,8 +113,8 @@ player={
       berry_timer, berry_count=0, 0
     end
 
-    for f in all(fruitrain) do
-      if f.type==fruit and not f.golden and berry_timer>5 and f then
+    for i,f in inext,fruitrain do
+      if f.type==fruit and not f.golden and berry_timer>5 then
         -- to be implemented:
         -- save berry
         -- save golden
@@ -130,7 +124,7 @@ player={
         init_object(lifeup, f.x, f.y,berry_count)
         del(fruitrain, f)
         destroy_object(f);
-        (fruitrain[1] or {}).target=_ENV
+        (fruitrain[i] or {}).target=f.target
       end
     end
     -- </fruitrain> --
@@ -146,21 +140,19 @@ player={
 
     -- jump buffer
     if jump then
-      jbuffer=4
-    elseif jbuffer>0 then
-      jbuffer-=1
+      jbuffer=5
     end
+    jbuffer=max(jbuffer-1)
 
     -- grace frames and dash restoration
     if on_ground then
-      grace=6
+      grace=7
       if djump<max_djump then
-        psfx(22)
+        psfx"22"
         djump=max_djump
       end
-    elseif grace>0 then
-      grace-=1
     end
+    grace=max(grace-1)
 
     -- dash effect timer (for dash-triggered events, e.g., berry blocks)
     dash_effect_time-=1
@@ -208,14 +200,14 @@ player={
       if jbuffer>0 then
         if grace>0 then
           -- normal jump
-          psfx(18)
+          psfx"18"
           jbuffer,grace,spd.y=0,0,-2
           init_smoke(0,4)
         else
           -- wall jump
-          local wall_dir=(is_solid(-3,0) and -1 or is_solid(3,0) and 1 or 0)
-          if wall_dir~=0 then
-            psfx(19)
+          local wall_dir=is_solid(-3,0) and -1 or is_solid(3,0) and 1
+          if wall_dir then
+            psfx"19"
             jbuffer,spd=0,vector(wall_dir*-2,-2)
             -- wall jump smoke
             init_smoke(wall_dir*6)
@@ -224,11 +216,10 @@ player={
       end
 
       -- dash
-      local d_full, d_half = 5, 3.5355339059 -- 5 * sqrt(2)
-
-      --<theo_crystal> --
-      if holding then
-        if dash then
+      -- <red_bubble> --
+      if dash or do_dash then
+        --<theo_crystal> --
+        if holding then
           -- throw
           if holding.is_solid(0,0) then -- can't throw
             psfx(21)
@@ -239,39 +230,38 @@ player={
             holding.cplayer=nil
             holding=nil
           end
-        end
-      -- <red_bubble> --
-      elseif djump>0 and dash or do_dash then
-      --</theo_crytal> --
-        do_dash=false
-      -- </red_bubble> --
-        init_smoke()
-        djump-=1
-        dash_time,_g.has_dashed,dash_effect_time=4, true, 10
-        -- vertical input
-        local v_input=btn(⬆️) and -1 or btn(⬇️) and 1 or 0
-        -- calculate dash speeds
-        spd=vector(h_input~=0 and
-        h_input*(v_input~=0 and d_half or d_full) or
-        (v_input~=0 and 0 or flip.x and -1 or 1)
-        ,v_input~=0 and v_input*(h_input~=0 and d_half or d_full) or 0)
-        -- effects
-        psfx(20)
-        _g.freeze,_g.shake=2,5
-        -- dash target speeds and accels
-        dash_target_x,dash_target_y,dash_accel_x,dash_accel_y=
-        2*sign(spd.x), (spd.y>=0 and 2 or 1.5)*sign(spd.y),
-        spd.y==0 and 1.5 or 1.06066017177 , spd.x==0 and 1.5 or 1.06066017177 -- 1.5 * sqrt()
+        elseif djump>0 then
+          --</theo_crystal>
+          do_dash=false
+          -- </red_bubble> --
+          init_smoke()
+          djump-=1
+          dash_time,_g.has_dashed,dash_effect_time=4, true, 10
+          -- vertical input
+          local v_input=btn(⬆️) and -1 or btn(⬇️) and 1 or 0
+          -- calculate dash speeds
+          local dspd=h_input&v_input==0 and 5 or 3.5355339059
+          spd=vector(h_input~=0 and h_input*dspd or
+          v_input~=0 and 0 or flip.x and -1 or 1,
+          v_input*dspd)
+          -- effects
+          psfx"20"
+          _g.freeze,_g.shake=2,5
+          -- dash target speeds and accels
+          dash_target_x,dash_target_y,dash_accel_x,dash_accel_y=
+          2*sign(spd.x), split"-1.5,0,2"[v_input+2],
+          v_input==0 and 1.5 or 1.06066017177 , spd.x==0 and 1.5 or 1.06066017177 -- 1.5 * sqrt()
 
-        -- emulate soft dashes
-        if h_input~=0 and ph_input==-h_input and oob(ph_input,0) then
-          spd.x=0
-        end
 
-      elseif djump<=0 and dash then
-        -- failed dash smoke
-        psfx(21)
-        init_smoke()
+          -- emulate soft dashes
+          if ph_input==-h_input and oob(ph_input,0) then
+            spd.x=0
+          end
+        else
+          -- failed dash smoke
+          psfx"21"
+          init_smoke()
+        end
       end
     end
 
@@ -323,16 +313,16 @@ end
 
 function update_hair(_ENV)
   local last=vector(x+(flip.x and 6 or 1),y+(btn(⬇️) and 4 or 2.9))
-  for h in all(hair) do
+  foreach(hair, function(h)
     h.x+=(last.x-h.x)/1.5
     h.y+=(last.y+0.5-h.y)/1.5
     last=h
-  end
+  end)
 end
 
 function draw_hair(_ENV)
-  for i,h in ipairs(hair) do
-    circfill(round(h.x),round(h.y),mid(4-i,1,2),8)
+  for i,h in inext,hair do
+    circfill(round(h.x),round(h.y),split"2,2,1,1,1"[i],8)
   end
 end
 
@@ -341,7 +331,7 @@ end
 player_spawn={
   init=function(_ENV)
     layer=2
-    sfx(15)
+    sfx"15"
     sprite=3
     target=y
     y=min(y+48,lvl_ph)
@@ -354,16 +344,17 @@ player_spawn={
     --- <keydoor> ---
     key_count=0
     --- <fruitrain> ---
-    for i=1,#fruitrain do
-      local f=init_object(fruitrain[i].type,x,y,fruitrain[i].sprite)
-      if fruitrain[i].type==key then
+    foreach(fruitrain, function(f)
+      --this gets called many times but saves tokens for checking if fruitrain is empty
+      fruitrain[1].target=_ENV
+
+      add(objects,f)
+      f.x,f.y=x,y
+      f.type.init(f,true)
+      if f.type==key then
         key_count+=1
       end
-      f.target=i==1 and _ENV or fruitrain[i-1]
-      f.r=fruitrain[i].r
-      f.fruit_id=fruitrain[i].fruit_id
-      fruitrain[i]=f
-    end
+    end)
     --- </fruitrain> </keydoor> ---
   end,
   update=function(_ENV)
@@ -382,7 +373,7 @@ player_spawn={
           -- clamp at target y
           y,spd,state,delay,_g.shake=target,vector(0,0),2,5,4
           init_smoke(0,4)
-          sfx(16)
+          sfx"16"
         end
       end
     -- landing and spawning player object
@@ -393,10 +384,10 @@ player_spawn={
         destroy_object(_ENV)
         local p=init_object(player,x,y)
         -- <keydoor> --
-        p.key_count=key_count
+        p.key_count=key_count;
         -- </keydoor> --
         --- <fruitrain> ---
-        if (fruitrain[1]) fruitrain[1].target=p
+        (fruitrain[1] or {}).target=p
         --- </fruitrain> ---
       end
     end
@@ -436,13 +427,18 @@ spring={
   end,
   update=function(_ENV)
     delta*=0.75
+    --can save tokens by setting hit as _ENV
+    --but i'm not desperate enough yet
     local hit=player_here()
     if hit then
-      hit.move(dir==0 and 0 or x+dir*4-hit.x,dir==0 and y-hit.y-4 or 0,1)
-      hit.spd=vector(
-      dir==0 and hit.spd.x*0.2 or dir*3,
-      dir==0 and -3 or -1.5
-      )
+      if dir==0 then
+        hit.move(0,y-hit.y-4,1)
+        hit.spd.x*=0.2
+        hit.spd.y=-3
+      else
+        hit.move(x+dir*4-hit.x,0,1)
+        hit.spd=vector(dir*3,-1.5)
+      end
       hit.dash_time,hit.dash_effect_time,delta,hit.djump=0,0,4,max_djump
     end
   end,
@@ -451,7 +447,7 @@ spring={
     if dir==0 then
       sspr(72,0,8,8-delta,x,y+delta)
     else
-      sspr(64,0,8-delta,8,dir==-1 and x+delta or x,y,8-delta,8,dir==1)
+      spr(8,dir==-1 and x+delta or x,y,1-delta/8,1,dir==1)
     end
   end
 }
@@ -464,14 +460,14 @@ refill={
     if timer>0 then
       timer-=1
       if timer==0 then
-        psfx(12)
+        psfx"12"
         init_smoke()
       end
     else
       offset+=0.02
       local hit=player_here()
       if hit and hit.djump<max_djump then
-        psfx(11)
+        psfx"11"
         init_smoke()
         hit.djump,timer=max_djump,60
       end
@@ -481,7 +477,7 @@ refill={
     if timer==0 then
       spr(15,x,y+sin(offset)+0.5)
     else
-      -- color(7)
+      -- color"7"
       -- line(x,y+4,x+3,y+7)
       -- line(x+4,y+7,x+7,y+4)
       -- line(x+7,y+3,x+4,y)
@@ -501,46 +497,44 @@ refill={
 
 fall_floor={
   init=function(_ENV)
-    solid_obj,state,unsafe_ground=true,0,true
+    solid_obj,state,unsafe_ground,delay=true,0,true,0
   end,
   update=function(_ENV)
+    --it looks like weird stuff goes on here with the decimal constants (mostly to ensure rounding correctly), but it should be equivalent to vanilla
+    --(and if i made an error, probably no one cares)
     -- idling
-    if state==0 then
-      for i=0,2 do
-        if check(player,i-1,-(i%2)) then
-          psfx(13)
-          state,delay=1,15
+    if delay>0 then
+      delay-=0.2
+    elseif state==0 then
+      for i=-1,1 do
+        if check(player,i,abs(i)-1) then
+          psfx"13"
+          state,delay=1,2.79
           init_smoke()
           break
         end
       end
     -- shaking
     elseif state==1 then
-      delay-=1
-      if delay<=0 then
-        state,delay,collideable=2,60--,false
-      end
+      state,delay,collideable=2,11.79--,false
     -- invisible, waiting to reset
     else
-      delay-=1
-      if delay<=0 and not player_here() then
-        psfx(12)
+      if not player_here() then
+        psfx"12"
         state,collideable=0,true
         init_smoke()
       end
     end
-  end,
-  draw=function(_ENV)
-    spr(state==1 and 26-delay/5 or state==0 and 23,x,y) --add an if statement if you use sprite 0
+    --if sprite 0 is not empty, need to fixup this
+    sprite=state==1 and 25.8-delay or state==0 and 23
   end
 }
 
 smoke={
   init=function(_ENV)
-    layer=3
-    spd,flip=vector(0.3+rnd(0.2),-0.1),vector(maybe(),maybe())
-    x+=-1+rnd(2)
-    y+=-1+rnd(2)
+    layer,spd,flip=3,vector(0.3+rnd"0.2",-0.1),vector(rnd()<0.5,rnd()<0.5)
+    x+=-1+rnd"2"
+    y+=-1+rnd"2"
   end,
   update=function(_ENV)
     sprite+=0.2
@@ -572,7 +566,7 @@ fruit={
       local hit=player_here()
       if hit then
         hit.berry_timer,target,r=
-        0,#fruitrain==0 and hit or fruitrain[#fruitrain],#fruitrain==0 and 12 or 8
+        0,fruitrain[#fruitrain] or hit,fruitrain[1] and 8 or 12
         add(fruitrain,_ENV)
       end
     end
@@ -593,7 +587,7 @@ fly_fruit={
       sfx_delay-=1
       if sfx_delay==0 then
        _g.sfx_timer=20
-       sfx(10)
+       sfx"10"
       end
       spd.y=appr(spd.y,-3.5,0.25)
       if y<-16 then
@@ -629,7 +623,7 @@ fly_fruit={
 lifeup={
   init=function(_ENV)
     spd.y,duration,flash,_g.sfx_timer,outline=-0.25,30,0,20--,false
-    sfx(9)
+    sfx"9"
   end,
   update=function(_ENV)
     duration-=1
@@ -640,14 +634,14 @@ lifeup={
   end,
   draw=function(_ENV)
     --<fruitrain>--
-    ?sprite<=5 and sprite.."000" or "1UP",x-4,y-4,7+flash%2
+    ?split"1000,2000,3000,4000,5000,1up"[min(sprite,6)],x-4,y-4,7+flash%2
     --<fruitrain>--
   end
 }
 key_door_used={}
 key={
-  init=function(_ENV)
-    if key_door_used[fruit_id] then
+  init=function(_ENV,exists)
+    if key_door_used[fruit_id] and not exists then
       destroy_object(_ENV)
     end
     y_=y
@@ -1307,11 +1301,10 @@ function init_object(_type,sx,sy,tile)
     return
   end
   --local _g=_g
-  local _ENV={}
+  local _ENV=setmetatable({},{__index=_g})
   type, collideable, sprite, flip, x, y, hitbox, spd, rem, fruit_id, outline, draw_seed=
-  _type, true, tile, _g.vector(), sx, sy, _g.rectangle(0,0,8,8), _g.vector(0,0), _g.vector(0,0), id, true, _g.rnd()
+  _type, true, tile, vector(), sx, sy, rectangle(0,0,8,8), vector(0,0), vector(0,0), id, true, rnd()
 
-  _g.setmetatable(_ENV,{__index=_g})
   function left() return x+hitbox.x end
   function right() return left()+hitbox.w-1 end
   function top() return y+hitbox.y end
@@ -1373,7 +1366,7 @@ function init_object(_type,sx,sy,tile)
     end
     -- </theo_crystal> --
     for axis in all{"x","y"} do
-      rem[axis]+=axis=="x" and ox or oy
+      rem[axis]+=vector(ox,oy)[axis]
       local amt=round(rem[axis])
       rem[axis]-=amt
 
@@ -1383,11 +1376,11 @@ function init_object(_type,sx,sy,tile)
         local step,p=sign(amt),_ENV[axis]
         local d=axis=="x" and step or 0
         for i=start,abs(amt) do
-          if not (is_solid(d,step-d) or oob(d,step-d)) then
-            _ENV[axis]+=step
-          else
+          if is_solid(d,step-d) or oob(d,step-d) then
             spd[axis],rem[axis]=0,0
             break
+          else
+            _ENV[axis]+=step
           end
         end
         movamt=_ENV[axis]-p --save how many px moved to use later for solids
@@ -1416,7 +1409,7 @@ function init_object(_type,sx,sy,tile)
             kill_player(hit)
           end
         elseif riding then
-          riding.move(axis=="x" and movamt or 0, axis=="y" and movamt or 0,1)
+          riding.move(vector(movamt,0)[axis],vector(0,movamt)[axis],1)
         end
         collideable=true
       end
@@ -1441,10 +1434,9 @@ function destroy_object(obj)
 end
 
 function kill_player(obj)
-  sfx_timer=12
-  sfx(17)
+  sfx_timer,shake=12,9
+  sfx"17"
   deaths+=1
-  shake=9
   destroy_object(obj)
   --dead_particles={}
   for dir=0,0.875,0.125 do
@@ -1456,16 +1448,16 @@ function kill_player(obj)
       dy=cos(dir)*3
     })
   end
-    -- <fruitrain> ---
-  for f in all(fruitrain) do
-    if (f.golden) full_restart=true
-    -- <keydoor> --
-    if f.type==fruit then
-      del(fruitrain,f)
+  -- <fruitrain> <keydoor> --
+  local keys={}
+  foreach(fruitrain,function(f)
+    full_restart = full_restart or f.golden
+    if f.type==key then
+      add(keys,f)
     end
-    -- </keydoor> --
-  end
-  --- </fruitrain> ---
+  end)
+  fruitrain=keys
+  --- </fruitrain> </keydoor> ---
   --delay_restart=15
   -- <transition> --
   co_trans=cocreate(transition)
@@ -1502,7 +1494,7 @@ function load_level(id)
   local exits=tonum(tbl[5]) or 0b0001
 
   -- exit_top,exit_right,exit_bottom,exit_left=exits&1!=0,exits&2!=0,exits&4!=0, exits&8!=0
-  for i,v in ipairs(split"exit_top,exit_right,exit_bottom,exit_left") do
+  for i,v in inext,split"exit_top,exit_right,exit_bottom,exit_left" do
     _ENV[v]=exits&(0.5<<i)~=0
   end
 
@@ -1562,16 +1554,7 @@ function _update()
   end
   frames%=30
 
-  if music_timer>0 then
-    music_timer-=1
-    if music_timer<=0 then
-      music(10,0,7)
-    end
-  end
-
-  if sfx_timer>0 then
-    sfx_timer-=1
-  end
+  sfx_timer=max(sfx_timer-1)
 
   -- cancel if freeze
   if freeze>0 then
@@ -1611,7 +1594,6 @@ function _update()
   foreach(objects,function(_ENV)
     if type==player or type==player_spawn then
       move_camera(_ENV)
-      return
     end
   end)
 
@@ -1633,8 +1615,8 @@ function _draw()
   if shake>0 then
     shake-=1
     if screenshake then
-      draw_x+=-2+rnd(5)
-      draw_y+=-2+rnd(5)
+      draw_x+=-2+rnd"5"
+      draw_y+=-2+rnd"5"
     end
   end
   camera(draw_x,draw_y)
@@ -1647,7 +1629,7 @@ function _draw()
     x+=spd-_g.cam_spdx
     _g.rectfill(x+_g.draw_x,y+_g.draw_y,x+w+_g.draw_x,y+16-w*0.1875+_g.draw_y,1)
     if x>128 then
-      x,y=-w,_g.rnd(120)
+      x,y=-w,_g.rnd"120"
     end
   end)
 
@@ -1655,13 +1637,13 @@ function _draw()
   map(lvl_x,lvl_y,0,0,lvl_w,lvl_h,4)
 
   -- draw outlines
-  for i=0,15 do pal(i,1) end
+  pal(split"1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1")
   pal=time
   foreach(objects,function(_ENV)
     if outline then
-      for dx=-1,1 do for dy=-1,1 do if dx&dy==0 then
-        camera(draw_x+dx,draw_y+dy) draw_object(_ENV)
-      end end end
+      for i=1,4 do
+        camera(draw_x+split"-1,0,0,1"[i],draw_y+split"0,-1,1,0"[i]) draw_object(_ENV)
+      end
     end
   end)
   pal=_pal
@@ -1699,9 +1681,9 @@ function _draw()
     off+=_g.min(0.05,spd/32)
     _g.rectfill(x+_g.draw_x,y+_g.draw_y,x+s+_g.draw_x,y+s+_g.draw_y,c)
     if x>132 then
-      x,y=-4,_g.rnd128()
+      x,y=-4,_g.rnd"128"
     elseif x<-4 then
-      x,y=128,_g.rnd128()
+      x,y=128,_g.rnd"128"
     end
   end)
 
@@ -1748,7 +1730,7 @@ end
 
 
 function two_digit_str(x)
-  return x<10 and "0"..x or x
+  return sub("0"..x,-2)
 end
 
 -- [helper functions]
@@ -1767,10 +1749,6 @@ end
 
 function cube(n)
   return n*n*n
-end
-
-function maybe()
-  return rnd()<0.5
 end
 
 function tile_at(x,y)
